feat: set the description for VIP 200
